requirements.txt
discord.py==1.7.3
keep_alive==0.0.1
requests==2.28.1
